+++ test/Rebus.OperationsDB.Tests/Commands/NestedNestedSagaCommand.cs
@@ -1,0 +1,6 @@
+namespace Dbosoft.Rebus.OperationsDB.Tests.Commands;
+
+public class NestedNestedSagaCommand
+{
+
+}
