+++ test/Rebus.OperationsDB.Tests/Commands/NestedSagaCommand.cs
@@ -1,0 +1,6 @@
+namespace Dbosoft.Rebus.OperationsDB.Tests.Commands;
+
+public class NestedSagaCommand
+{
+
+}
