+++ test/Rebus.OperationsDB.Tests/Commands/SimpleCommand.cs
@@ -1,0 +1,6 @@
+namespace Dbosoft.Rebus.OperationsDB.Tests.Commands;
+
+public class SimpleCommand
+{
+
+}
