+++ test/Rebus.OperationsDB.Tests/Commands/SubCommand1.cs
@@ -1,0 +1,6 @@
+namespace Dbosoft.Rebus.OperationsDB.Tests.Commands;
+
+public class SubCommand1
+{
+
+}
