+++ test/Rebus.OperationsDB.Tests/Commands/SubCommand2.cs
@@ -1,0 +1,5 @@
+namespace Dbosoft.Rebus.OperationsDB.Tests.Commands;
+
+public class SubCommand2
+{
+}
