+++ test/Rebus.OperationsDB.Tests/Commands/SubCommand3.cs
@@ -1,0 +1,6 @@
+namespace Dbosoft.Rebus.OperationsDB.Tests.Commands;
+
+public class SubCommand3
+{
+
+}
