+++ test/Rebus.OperationsDB.Tests/DatabaseTests.cs
@@ -21,7 +21,7 @@ using Microsoft.EntityFrameworkCore.Diagnostics;
 using Rebus.TransactionScopes;
 using Xunit;
 using Xunit.Abstractions;
-using static System.Formats.Asn1.AsnWriter;
+using Dbosoft.Rebus.OperationsDB.Tests.Commands;
 
 namespace Dbosoft.Rebus.OperationsDB.Tests;
