+++ test/Rebus.OperationsDB.Tests/Handlers/NestedSagaData.cs
@@ -1,0 +1,10 @@
+using Dbosoft.Rebus.Operations.Workflow;
+
+namespace Dbosoft.Rebus.OperationsDB.Tests.Handlers;
+
+public class NestedSagaData : TaskWorkflowSagaData
+{
+    public bool SubCommand2Completed { get; set; }
+    public bool SubCommand3Completed { get; set; }
+    public bool SagaCompleted { get; set; }
+}
