+++ test/Rebus.OperationsDB.Tests/Handlers/SimpleCommandHandler.cs
@@ -1,8 +1,9 @@
 using Dbosoft.Rebus.Operations;
+using Dbosoft.Rebus.OperationsDB.Tests.Commands;
 using JetBrains.Annotations;
 using Rebus.Handlers;
 
-namespace Dbosoft.Rebus.OperationsDB.Tests;
+namespace Dbosoft.Rebus.OperationsDB.Tests.Handlers;
 
 [UsedImplicitly]
 public class SimpleCommandHandler : IHandleMessages<OperationTask<SimpleCommand>>
