--- test/Rebus.OperationsDB.Tests/NestedNestedSagaCommand.cs
@@ -1,6 +1,0 @@
-namespace Dbosoft.Rebus.OperationsDB.Tests;
-
-public class NestedNestedSagaCommand
-{
-        
-}
