--- test/Rebus.OperationsDB.Tests/NestedSagaCommand.cs
@@ -1,6 +1,0 @@
-namespace Dbosoft.Rebus.OperationsDB.Tests;
-
-public class NestedSagaCommand
-{
-        
-}
