--- test/Rebus.OperationsDB.Tests/SubCommand1.cs
@@ -1,6 +1,0 @@
-namespace Dbosoft.Rebus.OperationsDB.Tests;
-
-public class SubCommand1
-{
-        
-}
