--- test/Rebus.OperationsDB.Tests/SubCommand2.cs
@@ -1,5 +1,0 @@
-namespace Dbosoft.Rebus.OperationsDB.Tests;
-
-public class SubCommand2
-{
-}
